Build: Fix compilation warning on array subscript of stack seal region
The stack seal region is defined as an 8 byte region in the linker
scripts, then declared as an extern uint32_t i.e 4 bytes value.

The compiler generates a warning when this is cast to a uint64_t *
in order to write the 8 byte stack seal value.

Correctly declare the stack seal region as an 8 byte region.

The CMSIS stack seal function accepts an uint32_t pointer then
itself cast it back to a uint64_t pointer.

+++ secure_fw/spm/include/tfm_arch_v8m.h
@@ -60,7 +60,7 @@
 /* Enable NS exceptions by setting NS PRIMASK to 0 */
 #define TFM_NS_EXC_ENABLE()     __TZ_set_PRIMASK_NS(0)
 
-extern uint32_t __STACK_SEAL;
+extern uint64_t __STACK_SEAL;
 
 /**
  * \brief Check whether Secure or Non-secure stack is used to restore stack
